still sinking?

+++ got/.i3/low_battery.sh
@@ -1,6 +1,17 @@
 #!/bin/bash
 
-mpv ~/.i3/sinking.mp3
+while true; do
+   battery=$(cat /sys/class/power_supply/BAT0/capacity)
 
-sleep 180
-systemctl poweroff
+   if [ "$battery" -le "5" ]; then
+      mpv ~/.i3/sinking.mp3
+      sleep 180
+      battery_update=$(cat /sys/class/power_supply/BAT0/capacity)
+      if [ "$battery_update" -ge "5" ]; then
+         continue
+      else
+         systemctl poweroff
+      fi
+   fi
+   sleep 30
+done
